requirements.txt
Flask == 2.0.2
numpy == 1.20.3
onvif_zeep == 0.2.12
requests == 2.26.0
